% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccode.R
\name{ccode}
\alias{ccode}
\title{Converting country codes}
\usage{
ccode(
  sourcevar,
  origin = NULL,
  destination = defaultcountrycode(),
  warn = TRUE,
  custom_dict = NULL,
  custom_match = NULL,
  origin_regex = FALSE,
  leaveifNA = FALSE,
  strict = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{sourcevar}{vector which contains the codes or country names to be converted (character or numeric)}

\item{origin}{Coding scheme of origin (case-insensitive string enclosed in quotes ""):  'iso2m', 'iso2c','iso3c', 'ecb', 'ec', 'name'(=name.en), 'name.fr', 'name.de', 'iso3n', 'un', 'wb', 'imf', 'fao', 'fips105', 'icao', 'ioc', 'cowc', 'ameco', 'oecd', 'wits', 'ccy'
If left unspecified, then the function will guess whether the source convention is one of 'iso2m', 'iso3c', 'name' or 'iso3n'}

\item{destination}{Coding scheme of destination (case-insensitive string enclosed in quotes ""): all of the above, plus 'name.en_estat',  'name.en_ecb', 'continent', 'region', 'sovereign'}

\item{warn}{Warns if elements from sourcevar could not be matched}

\item{custom_dict}{A data frame which supplies custom country codes.
Variables correspond to country codes, observations must refer to unique countries. See \code{countrycode_dict()$get()} for an example}

\item{custom_match}{retained for compatibility with countrycode package. Not operational at the moment.}

\item{origin_regex}{retained for compatibility with countrycode package. Not operational at the moment.}

\item{leaveifNA}{if FALSE then matches that cannot be converted are signalled as \code{NA}.
If TRUE, then  matches that cannot be converted are left as provided in \code{sourcevar}}

\item{strict}{Boolean. If FALSE, then 'old' codes are converted to new ones in a fuzzy manner. \code{strict=TRUE} applies strict matching to entities, and ignores 'unoffical' codes. See Disclaimer for details.}

\item{verbose}{Boolean. If TRUE, then the function posts messages ans warnings about things that look odd}
}
\value{
a string vector with converted country codes/names
}
\description{
Converts country codes (e.g. 3-letter to 2-letter), as well as from and to country names.
Strongly inspired by the \code{\link[countrycode]{countrycode}} package
}
\details{
There are four types of entities: existing countries and territories, former political entities (such as Yugoslavia), inter- and supra-national institutions and organisations (such as the IMF), and country groups (such as the euro area).

 The following code tables are available (for \code{origin} and \code{destination})
\itemize{
  \item 2 (to and from): same as\code{iso2m} below
  \item 3 (to and from): same as\code{iso3c} below
  \item iso2m (to and from): ISO 2-letter country codes (\code{FR} for France). Complemented with Estat memnonics for country groups (such as \code{EU} for the EU) and institutions (such as \code{EIB} for the EIB)
  \item iso2c (to and from): ISO 2-letter country codes (\code{FR} for France). Complemented with ECB codes for country groups (format LETTER-DIGIT) and institutions (format DIGIT-LETTER)
  \item iso3c (to and from): ISO 3-letter country codes (\code{FRA} for France). Complemented with ECB codes for country groups (format LETTER-DIGIT such as D0 for the EU) and institutions (format DIGIT-LETTER such as 4C for the EIB)
  \item ecb (to and from): ECB convention very similar to \code{iso2c} (such as \code{FR} for France). Some peculiarities for entities that have ceased to exist (e.g. Yugoslavia). See \href{https://sdw.ecb.europa.eu/datastructure.do?conceptMnemonic=REF_AREA&datasetinstanceid=143}{ECB codelist}.
  \item ec (to and from, alias eu, eurostat, com, commission): European Commission convention very similar to \code{iso2m}, i.e. ISO 2-letter country codes (\code{FR} for France) with an exception for Greece ('EL' instead of 'GR') and the UK ('UK' instead of 'GB'). See \href{https://webgate.ec.europa.eu/sdmxregistry/}{SDMX codelist registry}
  \item sovereign (to): The existing sovereign state to which a territory belongs, as recognized by the majority of UN Member States (i.e, Britain, Jersey, and Bermuda all yield 'UK' as sovereign). For former entities, sovereign denotes the successor with the largest population (e.g. Russia in the case of the USSR)
  \item name.en (to and from, alias 'name' and 'country.name'): short English language name. Uses regular expression to convert from names (e.g. Belgium to BE)
  \item name.fr (to, alias country.name.fr): French-language name, as defined by Eurostat
  \item name.de (to and from,, alias country.name.de): German-language name, as defined by Eurostat. also uses regular taken from the countrycode package
  \item name.en_estat (to): English language name as defined by Eurostat
  \item name.en_ecb (to): English language name as defined by ECB
  \item continent (to): Continent of a territory. Reads 'Organisation' for entities like 'IMF' and 'CountryGroup' for entities like 'EU'
  \item region (to): continental sub-region of a territory, as defined by the Worldbank. Type of organization or group (International or European) for institutions and country groups
  \item iso3n (to and from): ISO 3-digit numerical code
  \item un (to and from): UN numerical code (similar to iso3n, but with agggregates). See \href{https://unctadstat.unctad.org/en/classifications.html}{UNCTAD codelist}.
  \item wb (to and from): Worldbank 3-letter code (variant of iso3c). See \href{https://wits.worldbank.org/wits/wits/witshelp/content/codes/country_codes.htm}{Worldbank codelist}.
  \item wvs (to and from): World Values Survey numeric code (similar to iso3n). See \href{https://github.com/xmarquez/WorldValuesSurvey/blob/master/Country codes for WVS wave 6.csv}{WVS codelist}.
  \item fao (to and from): FAO numeric code. See \href{http://www.fao.org/countryprofiles/iso3list/en/}{FAO codelist}
  \item fips105 (to and from): FIPS 10-4 (Federal Information Processing Standard), similar to iso2c
  \item icao (to and from): ICAO (int'l civil air organization) code (when to: only first one relevant for a country is shown, when from: matches to largest population country within the ICAO area)
  \item imf (to and from): IMF numeric code
  \item ioc (to and from, alias olympic, sport): International Olympic Committee 3-letter code
  \item cowc (to and from): Correlates of War character code
  \item cown (to and from): Correlates of War numeric code
  \item ameco (to and from): AMECO database 3-letter code (variant of iso3c)
  \item oecd (to and from): OECD 3-letter code (variant of iso3c)
  \item ilo (to and from): Int'l Labor Organization 3-letter code (variant of iso3c)
  \item wits (to and from): 3-letter codes from Worldbank's World Integrated Trade Solution (variant of iso3c)
  \item ccy (to and from): 3-letter currency codes according to ISO 4217, such as \code{JPY} for Japan. Note that old currency codes such as FRF for French Franc are also matched.
}
Note that \code{2} is shorthand for \code{iso2m} and \code{3} is shorthand for \code{iso3c}
}
\section{Disclaimer}{


Note that there is absolutely no guarantee that the codes contained herein  conform to official code lists by the mentioned institutions. The code lists may not necessarily represent the views of
any institution regarding the status of territories included in, or excluded from, this package.
The code tables aim to represent the sovereignty status of entities as recognized by the majority of UN Member States, but accuracy is not guaranteed.
For creating your own deviations, adjustments, and additions to the correspondence table, see \code{\link{countrycode_dict}}.

The parameter \code{strict} tries to emulate official correspondence tables (accuracy not guaranteed): If strict=TRUE, certain entities are not matched under correspondence tables (e.g. Kosovo for \code{un}).
In addition, with strict=TRUE, old codes are mostly matched to old codes: \code{iso3n} code \code{278} for Western Germany is matched to Western Germany, rather than to modern Germany (which has code \code{280}).
In contrast \code{strict=FALSE} fills in unofficial codelist extensions for disputed territories, and matches old to new codes where the continuation is clear (e.g. codes for pre-1991 Ethiopia (incl. Eritrea) are matched to modern Ethiopia.)
}

\section{Relationship to countrycode package}{


This function is inspired by Vincent Arel-Bundock's \code{countrycode}  package.
The arguments of \code{ccode} are made to be consistent with those of the packages \code{countrycode}  function.
This function can be considered a fork of the countrycode packages, putting more emphasis on macroeconomic aspects, in particular:
\itemize{
\item Capturing small territories, also by regex
\item Adding institutions and country group codes
\item Adding alternative code matching
\item Reflecting the latest name and code changes etc. (e.g. "North Macedonia")
}
}

\examples{
ccode('Belgium','name','iso2m')

x1= c("France","Guadeloupe","Latvia","Guinea-Bissau","Euro area 19","Greece")
x2=ccode(x1,'name','iso2c')
print(x2)
x3=ccode(x2,'iso2c','ec') #note the code change for euro area and Greece
print(x3)


ccode(x2,'iso2c','sovereign', warn=FALSE) #Note the code change for Guadeloupe. There is no sovereign country to which the euro area belongs.
ccode(x2,'iso2c','sov', warn=FALSE, leaveifNA=TRUE) #leave the euro area code intact

ccode(x3,'iso2c','iso2c') #even though Greece was named 'EL' rather than 'GR', ccode can deal with that


ccode('CA','iso2m','ccy') #Canada has Canadian dollars


ccode('CA',2,'ccy') #2 is shorthand for iso2m
ccode(c('CA','MX'),2,3) #3 is shorthand for iso3c


ccode(cgrp('EU'),2,'name') #EU member names

ccode(c('European Commission','ECB','IMF'),'name',2) #use the of some institutions to find their codes
ccode(c('R4','R5','R6'),2,'name')

}
\seealso{
\code{\link{cgrp}} for getting country group members,
 \code{\link{defaultcountrycode}} for setting a default destination,
 \code{\link{countrycode_dict}} for altering the conversion table,
 \code{\link[countrycode]{countrycode}} for the original inspiration
}
