% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccode.R
\name{cgrp}
\alias{cgrp}
\title{Get a vector of country group (e.g. euro area) members}
\usage{
cgrp(
  sgroup = c("EA", "EU", "OECD", "EFTA", "ASEAN", "G20", "OPEC", "G7", "NATO"),
  time = NULL,
  ccodeas = defaultcountrycode(),
  fail = TRUE
)
}
\arguments{
\item{sgroup}{the code of the group in Eurostat or ECB terms. E.g. 'EU', 'OECD' or 'EA19' (See Details). Note: sgroup is a singleton, thus only the first element of any vector is taken.}

\item{time}{group members at that specific  year (as integer, like 2012) or date (as date or string like '2017-05-23'). If left empty, then current composition is returned. Note: time needs to be a singleton (only its first element is taken).}

\item{ccodeas}{The format in which to return the members (such as \code{'iso2m'}, \code{'iso3c'} or \code{'ec'}). If left empty, then uses \code{\link{defaultcountrycode}}, resp. 'iso2m'}

\item{fail}{if TRUE, requesting a non-existing sgroup generates an error. If FALSE, this would return a zero-length character instead}
}
\value{
a character vector with the country codes of group members currently (if \code{time} is unspecified), respectively with members as of the end of the period specfied in \code{time}
}
\description{
Get a vector of country group (e.g. euro area) members
}
\details{
See \code{\link{ccode}} for permissible values for ccode.
cgrp currently has variants for the following groups: euro area (EA, alias U2), European Union (EU, alias B0), OECD (P0), EFTA (R2), OPEC (R5), G7 (R23), NATO
The following shortcuts can also be used: EA11, EA12, EA17, EA18, EA19, EU12, EU15, EU25, EU27, EU28 respectively their ECB equivalents (I1, I2, I6, I7, I8, B1, B2, B3, B6, B5)
}
\examples{
cgrp("EA") #euro area members
cgrp(EA)   #also works without quotes
cgrp("EA", 2002) #euro area Members in 2012
cgrp('U2', 2002) #same, using ECB instead of Estat code for euro area
cgrp(G20, ccodeas=3) #3 letter codes for G20 countries (B0 is EU)

cgrp('EA12') #shortcut: euro area with 12 Members
cgrp('EU','2004-01-31') #EU Members at the start of 2004 (excludes 'new' member states)

cgrp('EFTA', 1993)
cgrp('EFTA', 1995) #three countries had left EFTA as of 1995


cgrp('EFTA', 1994, "name")

cgrp('OECD', 2000, ccodeas="iso3c") #OECD countries in 2000
setdiff(cgrp('NATO',2010), cgrp('EU',2010)) #NATO countries outside the EU in 2010
cgrp(OPEC, 1974, ccodeas='name') #names of OPEC countries in 1974

}
\seealso{
\code{\link{ccode}} for converting between country code conventions, \code{\link{countrycode_dict}} to set and change your own country groups
}
