% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccode.R
\name{defaultcountrycode}
\alias{defaultcountrycode}
\title{Set the default country code to be converted to}
\usage{
defaultcountrycode(x)
}
\arguments{
\item{x}{the default 'destination' argument to function \code{ccode}.}
}
\value{
if x is not specified, then the function returns the current default country code. If x is specfied, the function sets the countrycode to x (and returns it invisibly)
}
\description{
Set the default country code to be converted to
}
\details{
See \code{\link{ccode}} for permissible values for x. Commonly used values are:
\itemize{
\item \code{"iso2c"} for 2-letter country codes ('FR' for France), with letter-digit combinations for country groups  (\code{U2} for Euro area)
\item \code{"iso2m"} for 2-letter country codes ('FR' for France), with mnemnonics for country groups  (\code{EA19} for Euro area 19)
\item \code{"ec"} for Eurostat/EC-style two-letter codes (EL instead of GR, UK instead of GB, otherwise like \code{iso2c}) and mnemonics for country groups (\code{EA} for Euro area)
\item \code{"iso3c"} for 3-letter country codes ('FRA' for France), with letter-digit combinations for country groups  (\code{U2} for Euro area)
}
}
\examples{
defaultcountrycode("iso2m") #set default country code to ISO 2-letters

ccode("Sudan","name") # here 'destination' does not have to be specfied in ccode
ccode("FRA","iso3c") # similar

defaultcountrycode() #shows current default country code

defaultcountrycode(NULL) # unsets default countrycode

}
\seealso{
\code{\link{ccode}}, \code{\link{cgrp}}
}
