% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictccode}
\alias{dictccode}
\title{Country code conversion tables}
\format{
An object of class \code{"data.frame"}. See \code{\link{countrycode_dict}} for explanations
}
\source{
\href{https://sdw.ecb.europa.eu/datastructure.do?conceptMnemonic=REF_AREA&datasetinstanceid=143}{ECB codelist},
\href{https://unctadstat.unctad.org/en/classifications.html}{UNCTAD codelist},
\href{http://www.fao.org/countryprofiles/iso3list/en/}{FAO codelist},
\href{https://github.com/xmarquez/WorldValuesSurvey/blob/master/Country codes for WVS wave 6.csv}{WVS codelist},
\href{https://wits.worldbank.org/wits/wits/witshelp/content/codes/country_codes.htm}{Worldbank codelist}
}
\usage{
data(dictccode)
}
\description{
Codes for territories, institutions, and country groups, including official names
}
\examples{
data(dictccode)
}
\keyword{datasets}
